requirements.txt
Flask
flask-cors
Werkzeug
openai
python-docx
PyPDF2
moviepy
tiktoken
langchain
weaviate-client